[492609] - Fixed bug where "added" module cache was out of synch
When error or cancellation occurs during a module publish, the module
cache that tracks a module that is being added in an operation was not
being cleaned up correctly, resulting in an out of date cache that would
prevent corrent module update on refresh or deletion.

Also added test case to cover the fix.