[498849] - Improved application delegate framework
Application delegate framework now allows fallback to lower priority
delegates when generating archives for an application, in the event that
a higher priority delegate is unable to archive the application. Also
cleaned up code so that there are distinct archivers for different
types: file from manifest.yml, war, and jar archiving.