[488784] - Runstate of app is now shown in editor
The servers view label provider is a decorator and doesn't seem to be
reusable in the editors TableViewer. Therefore refactored deployment
label code into reusable form that can be used in both Servers view and
editor. Also introduced new run state API in
CloudFoundryApplicationModule so that the editor also shows "starting"
state.